feat: add types to registration page
