+++ src/pages/registration/index.ts
