+++ src/pages/registration/interface.ts
@@ -1,0 +1,5 @@
+export interface IRegistrationPageState {
+  name: string;
+  email: string;
+  password: string;
+}
